rename network->indicators
