+++ snapcraft/internal/indicators.py
